+++ components/containers/DynamicForm/Settings.tsx
@@ -4,8 +4,11 @@ import { useTranslation } from "next-i18next";
 import { DeleteButton } from "../../forms/Button/DeleteButton";
 import { useRouter } from "next/router";
 import axios from "axios";
-import { logMessage } from "../../../lib/logger";
-import { FormDBConfigProperties } from "../../../lib/types";
+import { logMessage } from "@lib/logger";
+import { FormDBConfigProperties } from "@lib/types";
+import { Tab, Tabs, TabList, TabPanel } from "react-tabs";
+import "react-tabs/style/react-tabs.css";
+import BearerRefresh from "@components/admin/BearerRefresh/BearerRefresh";
 
 interface FormSettingsProps {
   form: FormDBConfigProperties;
@@ -40,7 +43,6 @@ export const FormSettings = (props: FormSettingsProps): React.ReactElement => {
   const { form } = props;
   const router = useRouter();
   const { t } = useTranslation("admin-templates");
-
   const newText =
     router.query && router.query.newForm ? (
       <p className="gc-confirmation-banner">{t("settings.new")}</p>
@@ -51,20 +53,32 @@ export const FormSettings = (props: FormSettingsProps): React.ReactElement => {
   return (
     <>
       <h1 className="gc-h1">{t("settings.title")}</h1>
-      <div>{newText}</div>
       <div data-testid="formID" className="mb-4">
         Form ID: {form.formID}
       </div>
-      <h2>Edit Form Config File:</h2>
-      <JSONUpload form={form}></JSONUpload>
-      <br />
-      <div>
-        <DeleteButton
-          action={handleDelete}
-          data={form.formID}
-          redirect={`/admin/view-templates`}
-        ></DeleteButton>
-      </div>
+      <Tabs>
+        <TabList>
+          <Tab>{t("settings.tabLabels.jsonUpload")}</Tab>
+          <Tab>{t("settings.tabLabels.bearerToken")}</Tab>
+        </TabList>
+
+        <TabPanel>
+          <div>{newText}</div>
+          <h2>{t("settings.edit")}</h2>
+          <JSONUpload form={form}></JSONUpload>
+          <br />
+          <div>
+            <DeleteButton
+              action={handleDelete}
+              data={form.formID}
+              redirect={`/admin/view-templates`}
+            ></DeleteButton>
+          </div>
+        </TabPanel>
+        <TabPanel>
+          <BearerRefresh formID={form.formID}></BearerRefresh>
+        </TabPanel>
+      </Tabs>
     </>
   );
 };